+++ src/click/core.py
@@ -413,6 +413,27 @@ class Context:
         self._source_by_paramname = {}
         self._exit_stack = ExitStack()
 
+    def to_info_dict(self):
+        """Gather information that could be useful for a tool generating
+        user-facing documentation. This traverses the entire CLI
+        structure.
+
+        .. code-block:: python
+
+            with Context(cli) as ctx:
+                info = ctx.to_info_dict()
+
+        .. versionadded:: 8.0
+        """
+        return {
+            "command": self.command.to_info_dict(self),
+            "info_name": self.info_name,
+            "allow_extra_args": self.allow_extra_args,
+            "allow_interspersed_args": self.allow_interspersed_args,
+            "ignore_unknown_options": self.ignore_unknown_options,
+            "auto_envvar_prefix": self.auto_envvar_prefix,
+        }
+
     def __enter__(self):
         self._depth += 1
         push_context(self)
@@ -632,6 +653,14 @@ class Context:
         """
         return self.command.get_help(self)
 
+    def _make_sub_context(self, command):
+        """Create a new context of the same type as this context, but
+        for a new command.
+
+        :meta private:
+        """
+        return type(self)(command, info_name=command.name, parent=self)
+
     def invoke(*args, **kwargs):  # noqa: B902
         """Invokes a command callback in exactly the way it expects.  There
         are two ways to invoke this method:
@@ -663,8 +692,7 @@ class Context:
                     "The given command does not have a callback that can be invoked."
                 )
 
-            # Create a new context of the same type as this context.
-            ctx = type(self)(other_cmd, info_name=other_cmd.name, parent=self)
+            ctx = self._make_sub_context(other_cmd)
 
             for param in other_cmd.params:
                 if param.name not in kwargs and param.expose_value:
@@ -767,6 +795,17 @@ class BaseCommand:
         self.context_settings = context_settings
 
     def to_info_dict(self, ctx):
+        """Gather information that could be useful for a tool generating
+        user-facing documentation. This traverses the entire structure
+        below this command.
+
+        Use :meth:`click.Context.to_info_dict` to traverse the entire
+        CLI structure.
+
+        :param ctx: A :class:`Context` representing this command.
+
+        .. versionadded:: 8.0
+        """
         return {"name": self.name}
 
     def __repr__(self):
@@ -1249,13 +1288,16 @@ class MultiCommand(Command):
 
     def to_info_dict(self, ctx):
         info_dict = super().to_info_dict(ctx)
-        info_dict.update(
-            commands=[
-                self.get_command(ctx, cmd_name).to_info_dict(ctx)
-                for cmd_name in self.list_commands(ctx)
-            ],
-            chain=self.chain,
-        )
+        commands = {}
+
+        for name in self.list_commands(ctx):
+            command = self.get_command(ctx, name)
+            sub_ctx = ctx._make_sub_context(command)
+
+            with sub_ctx.scope(cleanup=False):
+                commands[name] = command.to_info_dict(sub_ctx)
+
+        info_dict.update(commands=commands, chain=self.chain)
         return info_dict
 
     def collect_usage_pieces(self, ctx):
@@ -1688,6 +1730,14 @@ class Parameter:
         self.autocompletion = autocompletion
 
     def to_info_dict(self):
+        """Gather information that could be useful for a tool generating
+        user-facing documentation.
+
+        Use :meth:`click.Context.to_info_dict` to traverse the entire
+        CLI structure.
+
+        .. versionadded:: 8.0
+        """
         return {
             "name": self.name,
             "param_type_name": self.param_type_name,